refactor: Integrate ApiError for improved error handling in material services - Updated the  file to include the  class for standardized error handling. - Replaced generic errors with specific  instances for better clarity and debugging. - Enhanced error messages to provide more context, including error codes and descriptions. - Checked for material existence before updating or deleting, and provided detailed error messages if the material does not exist.
These changes improve error handling within the material services, providing more meaningful and consistent error messages, which aid in debugging and enhance the overall robustness of the application.

+++ src/services/materialServices.js
@@ -1,12 +1,13 @@
 const Dao = require('../dao')
 const Validations = require('../validations')
+const ApiError = require('../utils/ApiError')
 
 module.exports = {
     getAllMaterials: async () => {
         try {
             return (materials = await Dao.materialsDao.getAll())
         } catch (err) {
-            throw err
+            throw new ApiError('PRINTING_MATERIALS_0001', 500, 'Failed to fetch printing materials')
         }
     },
 
@@ -14,7 +15,11 @@ module.exports = {
         try {
             return (materials = await Dao.materialsDao.getById(id))
         } catch (err) {
-            throw err
+            throw new ApiError(
+                'PRINTING_MATERIALS_0002',
+                500,
+                `Failed to fetch a printing materials with id ${id}`,
+            )
         }
     },
 
@@ -24,25 +29,57 @@ module.exports = {
         try {
             return (materials = await Dao.materialsDao.create(requestData))
         } catch (err) {
-            throw err
+            throw new ApiError(
+                'PRINTING_MATERIALS_0003',
+                500,
+                `Error creating printing material with name ${requestData.name}`,
+            )
         }
     },
 
     updateMaterial: async (id, requestData) => {
+        // check if material exists
+        material = materials = await Dao.materialsDao.getById(id)
+        if (!material)
+            throw new ApiError(
+                'PRINTING_MATERIALS_0006',
+                404,
+                `3D Printing Material with ${id} does not exists`,
+            )
+
+        // validating the update request data
         const { error } = Validations.material.updateMaterial(requestData)
         if (error) throw error
+
         try {
             return await Dao.materialsDao.update(id, requestData)
         } catch (err) {
-            throw err
+            throw new ApiError(
+                'PRINTING_MATERIALS_0004',
+                500,
+                `Error creating printing material with id ${id}`,
+            )
         }
     },
 
     deleteMaterialById: async (id) => {
         try {
+            // check if material exists
+            material = materials = await Dao.materialsDao.getById(id)
+            if (!material)
+                throw new ApiError(
+                    'PRINTING_MATERIALS_0006',
+                    404,
+                    `3D Printing Material with ${id} does not exists`,
+                )
+
             return await Dao.materialsDao.delete(id)
         } catch (err) {
-            throw err
+            throw new ApiError(
+                'PRINTING_MATERIALS_0005',
+                500,
+                `Failed to delete a printing materials with id ${id}`,
+            )
         }
     },
 }
